Span factoring
